@@ -1,5 +1,5 @@
 import math
-from typing import Union, Tuple
+from typing import Tuple
 
 from tinygrad import Tensor, nn, dtypes
 from tinygrad.helpers import prod, argfix
@@ -56,7 +56,7 @@ class EmbeddingBert(nn.Embedding):
     return (arange == idx).mul(vals).sum(2, acc_dtype=vals.dtype)
 
 class LayerNormBert:
-  def __init__(self, normalized_shape:Union[int, Tuple[int, ...]], eps:float=1e-12, elementwise_affine:bool=True):
+  def __init__(self, normalized_shape:int | Tuple[int, ...], eps:float=1e-12, elementwise_affine:bool=True):
     self.normalized_shape = (normalized_shape,) if isinstance(normalized_shape, int) else tuple(normalized_shape)
     self.axis, self.eps, self.elementwise_affine = tuple(-1-i for i in range(len(self.normalized_shape))), eps, elementwise_affine
     self.weight, self.bias = (Tensor.ones(*self.normalized_shape, dtype=dtypes.float32), Tensor.zeros(*self.normalized_shape, dtype=dtypes.float32)) if elementwise_affine else (None, None)
